Fix warnings and separate source files (#2)
* fixed warinings with ints and unsigned ints, and misleading indentations

* fixed warnings unused variable

* fixed warnings uninitialized variables

* changes in source files such that the header/source separation can be done with the Lzz tool

* added src folder with:
separated source and header files .h and .cpp
makefile such that one can compile with make or make full

This is good practice and will allow for better integration with other codes, such as OPAL, where MITHRA is integrated as a solver.
The separation of header files was done with Lzz library, a.k.a. Lazy C++ from http://www.lazycplusplus.com/index.html

and some hand editing of details

I have left the source directory with the old source files just in case there is some bug that I have missed in the new files, but now the edits and changes should be done in the src directory, with funciton definitions in the .h files, and implementations in the .cpp files.